wip(objects): rename building ns
