+++ Objects/Objects/Building/NonPlanarRoof.cs
@@ -2,10 +2,10 @@
 using System.Collections.Generic;
 using System.Text;
 using Speckle.Core.Models;
-using Objects.BuildingObject;
+using Objects.Building;
 using Objects.Definitions;
 
-namespace Objects.BuildingObject
+namespace Objects.Building
 {
   public class NonPlanarRoof : CurveBasedElement
   {
