+++ Objects/Objects/Building/Rebar.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Objects.BuildingObject.enums;
+using Objects.Building.enums;
 using Objects.Definitions;
 
-namespace Objects.BuildingObject
+namespace Objects.Building
 {
   public class Rebar :CurveBasedElement
   {
