+++ Objects/Objects/Building/Space.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Text;
 using Objects.Definitions;
 
-namespace Objects.BuildingObject
+namespace Objects.Building
 {
-  public class Station : PointBasedElement
+  public class Space : PointBasedElement
   {
     // to implement source app parameters interface from claire
   }
