+++ Objects/Objects/Building/Station.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Text;
 using Objects.Definitions;
 
-namespace Objects.BuildingObject
+namespace Objects.Building
 {
-  public class Space : PointBasedElement
+  public class Station : PointBasedElement
   {
     // to implement source app parameters interface from claire
   }
