+++ Objects/Objects/Definitions/GenericFormProperty.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using Objects.BuildingObject.enums;
+using Objects.Building.enums;
 using Speckle.Core.Models;
 
 namespace Objects.Definitions
